Added a nonlinear solver for finding L and d (and m)
In calibration.py, find_d_L uses fmin from scipy.optmize to minimize
chisquared for the lambda peaks compared to the calibration wavelength.
Fitting.py was modified to handle this better.  It now handles if there
are not equal numbers of up and down points gracefully.  It will return
an empty list if there are no peaks above the threshold.

ToDo:
        -Check my L and d answer against the Ar II line that is not in
the same free spectral range of my Th I line.
        -Start fitting lamp profiles convolved with a Lorentzian
instrument function to get a voight line shape to compare to my ring
sums.