Add error propagation for ring sum
Error propagation is now being calculated in core/ringsum.py by
calculating weighted means. The motivation for this was realizing the
variation in the number of pixels in a given r (read annulus) bin.

The photon counting statistics is hardcoded based on measurements taken
of the intergration sphere.

$\sigma = \sqrt{1.8 \rm{text}}$

Link for weighted mean:
https://en.wikipedia.org/wiki/Weighted_arithmetic_mean#Dealing_with_variance