Modify argon plasma solver
I'm still torn on how I want to do the argon plasma solver. I believe
the goal is to have the finesse determined by the calibration and to get
the valley counts above the read noise to avoid an offset. This would
allow a naive 3 paramter fit when the velocity doesn't matter.